+++ src/sqlalchemy_declarative_extensions/dialects/mysql/__init__.py
@@ -1,0 +1,29 @@
+from sqlalchemy_declarative_extensions.dialects.mysql.function import (
+    Function,
+    FunctionDataAccess,
+    FunctionSecurity,
+)
+from sqlalchemy_declarative_extensions.dialects.mysql.procedure import (
+    Procedure,
+    ProcedureSecurity,
+)
+from sqlalchemy_declarative_extensions.dialects.mysql.trigger import (
+    Trigger,
+    TriggerEvents,
+    TriggerOrder,
+    TriggerOrders,
+    TriggerTimes,
+)
+
+__all__ = [
+    "Function",
+    "FunctionDataAccess",
+    "FunctionSecurity",
+    "Procedure",
+    "ProcedureSecurity",
+    "Trigger",
+    "TriggerEvents",
+    "TriggerOrder",
+    "TriggerOrders",
+    "TriggerTimes",
+]