+++ src/sqlalchemy_declarative_extensions/trigger/base.py
@@ -20,14 +20,10 @@ class Trigger:
     def named(self, name: str):
         return replace(self, name=name)
 
-    def to_sql_create(self, replace=False):
+    def to_sql_create(self):
         raise NotImplementedError()
 
     def to_sql_update(self, connection: Connection):
-        assert connection.dialect.server_version_info
-        if connection.dialect.server_version_info >= (14, 0):
-            return [self.to_sql_create(replace=True)]
-
         return [
             self.to_sql_drop(),
             self.to_sql_create(),