+++ CHANGELOG.md
@@ -1,8 +1,18 @@
 # Changelog
 
+## 0.13
+
+- feat: Add support for MetaData.drop_all.
+- feat: Add basic support for functions and procedures to MySQL.
+
+## 0.12
+
+- feat: Add basic support for triggers to MySQL.
+
 ## 0.11
 
 ### 0.11.4
+
 - fix: Incorrect when condition paren stripping.
 
 ### 0.11.3